server/requirements.txt
uvicorn
fastapi
torch
torchvision
transformers
sentencepiece
sacremoses
min_dalle
pymysql
python-dotenv